policy: tune feerate estimator parameters
Changes the static parameters used in CBlockPolicyEstimator to
support the Dogecoin network, inspired by changes made in Bitcoin
Core 0.15, and field-tested to survive and recover from peak
inflow of transactions.

In addition to changing the parameters:
- documented each changed parameter and the rationale in-line
- changed the fee estimator tests to take into account the new
  configuration. Note that only 8 buckets can be tested on a set
  of 10, when confidence is set to 80%

1. Changes the decay rate of statistical history from a half-life
   of 346 blocks to a half-life of 60 blocks. (DEFAULT_DECAY)

This means that the weight of previous block feerate stats on the
current fee estimates becomes smaller, and more recent block
results weigh heavier, so that the estimates react faster to
temporary peaks.

Weight:

age | previous |   now
1h  |    0.887 | 0.498
2h  |    0.786 | 0.248
3h  |    0.697 | 0.124
4h  |    0.618 | 0.062
5h  |    0.548 | 0.031
7h  |    0.431 | 0.008
10h |    0.301 | 0.001
24h |    0.056 | 0.001

2. Changes the historical confidence level (the percentage of
   transactions included at a certain level) from 95% to 80%.
   (MIN_SUCCESS_PCT)

This lowers the successful inclusion rate because miners in the
Dogecoin ecosystem tune mempool policy for orphan prevention
more than transaction inclusion (because subsidy > fees). This
negatively impacts the precision of the estimate, but with 10%
of empty blocks and at least 2 orders of magnitude difference
between miner enforced minimum fee rates, higher precision will
cause extremely high feerate estimate, yet still have low
effectivity rates. The value was picked by taking Bitcoin Core's
half-confidence of 60% used for 24-block half-time, and standard
confidence of 85% and picking a workable value between those.

3. Decreases the number of transactions per bucket required to
   consider a feerate valid from 1 to 0.1 (SUFFICIENT_FEETXS)

This makes more buckets relevant while filtering for fees to
consider and offsets the more rapid half-life from point 1, as
we still want to consider a feerate included with a single
transaction 10 blocks ago, yet because of decay would not be
sufficient for inclusion (because effective relevance would
be 0.89 with the new parametrization). Value taken from Bitcoin
Core because it's been the default there for v0.15 until v25.0

4. Set the minimum and maximum feerates to be considered in
   line with default mempool policies: min 0.001 DOGE/kB and
   max 10 DOGE/kB (MIN_FEERATE, MAX_FEERATE)

This is copied 1-on-1 from an earlier attempt by Ross Nicoll that
needed further investigation.

5. Lowers the spacing inside feerate buckets from exponent
   1.1 to 1.05 (FEE_SPACING)

This increases the granularity of the buckets and is copied
directly from Bitcoin Core, where this has been the parameter
from v0.15 until v25.0